requirements.txt
altair
attrs
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
gitdb
GitPython
h11
idna
Jinja2
jsonschema
jsonschema-specifications
markdown-it-py
MarkupSafe
mdurl
narwhals
numpy
openai
outcome
packaging
pandas
pillow
protobuf
pyarrow
pycparser
pydeck
PyGetWindow
Pygments
PyMuPDF
PyRect
PySocks
python-dateutil
pytz
referencing
reportlab
rich
rpds-py
selenium
six
smmap
sniffio
sortedcontainers
soupsieve
streamlit
tenacity
toml
tornado
trio
trio-websocket
typing_extensions
tzdata
urllib3
watchdog
websocket-client
wsproto
playwright
PyPDF2